Refactor ingredientChoose page with custom hooks
