+++ src/components/pages/IngredientChoose.jsx
@@ -1,59 +1,27 @@
-import React, { useState, useRef } from 'react'
+import React from 'react'
+import useCharm from '../../hooks/useCharm'
 import Navbar from '../globals/Navbar'
+import Select from 'react-select'
 import IngredientInfo from '../globals/IngredientInfo'
 import MealPrep from '../globals/MealPrep'
-import Select from 'react-select'
-import { toast } from 'react-toastify'
-import useIngredients from '../../hooks/useIngredients'
+import useAddIngredient from '../../hooks/useAddIngredient'
 
 const IngredientChoose = () => {
-  const { selectedIngredient, selectOptions, ingredientChart, selectIngredientRef } = useIngredients()
+  const {
+    selectedIngredient,
+    selectOptions,
+    ingredientChart,
+    selectIngredientRef
+  } = useCharm()
 
-  const [quantity, setQuantity] = useState(100)
-  const [meals, setMeals] = useState([])
-  const [mealsTotalMacros, setMealsTotalMacros] = useState({
-    proteins: 0,
-    carbohydrates: 0,
-    fats: 0,
-    calories: 0
-  })
-
-  const quantityRef = useRef()
-
-  const handleIngredientCuantity = (e) => {
-    const quantityValue = e.target.value
-    setQuantity(quantityValue)
-  }
-
-  const handleAddIngredient = (e) => {
-    e.preventDefault()
-
-    const maxIngredients = 10
-
-    const newMeal = {
-      name: selectedIngredient.name,
-      calories: selectedIngredient.calories * quantity / 100,
-      proteins: selectedIngredient.proteins * quantity / 100,
-      carbohydrates: selectedIngredient.carbohydrates * quantity / 100,
-      fats: selectedIngredient.fats * quantity / 100,
-      quantity
-    }
-
-    if (meals.length === maxIngredients) {
-      toast.error('Can not add more ingredients', { autoClose: 1500, theme: 'dark' })
-    } else {
-      setMeals([...meals, newMeal])
-      setMealsTotalMacros(
-        {
-          proteins: mealsTotalMacros.proteins + newMeal.proteins,
-          carbohydrates: mealsTotalMacros.carbohydrates + newMeal.carbohydrates,
-          fats: mealsTotalMacros.fats + newMeal.fats,
-          calories: mealsTotalMacros.calories + newMeal.calories
-        }
-      )
-      toast.success('Ingredient added', { autoClose: 1500, theme: 'dark' })
-    }
-  }
+  const {
+    handleAddIngredient,
+    quantityRef,
+    handleIngredientCuantity,
+    quantity,
+    meals,
+    mealsTotalMacros
+  } = useAddIngredient()
 
   return (
     <>
@@ -80,11 +48,14 @@ const IngredientChoose = () => {
 
           <div className='ingredientsChoose__charm'>
             <h3 className='ingredientsChoose__charm__title'>% of macronutrients</h3>
-            <IngredientInfo
-              proteins={selectedIngredient.proteins}
-              carbohydrates={selectedIngredient.carbohydrates}
-              fats={selectedIngredient.fats}
-            />
+            {!selectedIngredient && <p>Loading...</p>}
+            {selectedIngredient &&
+               <IngredientInfo
+                proteins={selectedIngredient.proteins}
+                carbohydrates={selectedIngredient.carbohydrates}
+                fats={selectedIngredient.fats}
+               />
+            }
           </div>
         </section>
         {/* PLAN YOUR MEAL SECTION */}
